Separate Image Handling and Integrate Functionality
